+++ spawner/cmd/setup.go
@@ -1,11 +1,11 @@
 package cmd
 
 import (
-	spawner "github.com/threefoldtech/guardians_healthchecker/spawner/pkg/deployer"
+	spawner "github.com/threefoldtech/guardians_healthchecker/spawner/pkg/spawner"
 	"github.com/threefoldtech/tfgrid-sdk-go/grid-client/deployer"
 )
 
-func setup(conf spawner.Config, debug bool) (deployer.TFPluginClient, error) {
+func setup(conf spawner.Config) (deployer.TFPluginClient, error) {
 	mnemonic := conf.Mnemonic
 
 	opts := []deployer.PluginOpt{
@@ -14,9 +14,6 @@ func setup(conf spawner.Config, debug bool) (deployer.TFPluginClient, error) {
 		deployer.WithProxyURL(conf.GridEndpoints.Proxy),
 		deployer.WithRelayURL(conf.GridEndpoints.Relay),
 		deployer.WithSubstrateURL(conf.GridEndpoints.SubsrateURL),
-	}
-	if debug {
-		opts = append(opts, deployer.WithLogs())
 	}
 
 	return deployer.NewTFPluginClient(mnemonic, opts...)
